+++ upload/tasks/03_set_puppetenv.sh
@@ -5,7 +5,16 @@ set -e
 source $temp_env/include
 
 makechanges=false
-cluster_csv="$HOME/upload/tasks/puppet/manifests/extdata/$hostname_f.csv"
+puppet_master_ip_file="$HOME/upload/tasks/puppet/01_update.sh"
+grep -qFx "^real_master='$elastic_ip'" $puppet_master_ip_file || makechanges=true
+if $makechanges; then
+	sed -i "/^real_master/d" $puppet_master_ip_file
+	sed -i "3ireal_master='$elastic_ip';" $puppet_master_ip_file
+fi
+
+[ "$hostnum" != 1 ] && exit 0
+
+cluster_csv="$HOME/upload/tasks/puppet/manifests/extdata/common.csv"
 [ -e $cluster_csv ] && rm -rf $cluster_csv
 mongodb_repo="deb http://downloads-distro.mongodb.org/repo/ubuntu-upstart dist 10gen"
 
@@ -22,11 +31,4 @@ echo "torque_complie_args_extra," | tee -a $cluster_csv
 echo "torque_spool_dir," | tee -a $cluster_csv
 echo "torque_user_not_root,ubuntu" | tee -a $cluster_csv
 echo "hostname_s," | tee -a $cluster_csv
-
-puppet_master_ip_file="$HOME/upload/tasks/puppet/01_update.sh"
-grep -qFx "^real_master='$elastic_ip'" $puppet_master_ip_file || makechanges=true
-if $makechanges; then
-       sed -i "/^real_master/d" $puppet_master_ip_file
-       sed -i "3ireal_master='$elastic_ip';" $puppet_master_ip_file
-fi
